Refactor `_pytask.mark_utils`. (#235)
