+++ docs/source/changes.md
@@ -20,6 +20,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
   :class:`typing.NamedTuple` for better typing.
 - {pull}`232` moves the documentation to MyST.
 - {pull}`234` removes `MetaTask`. There is only {class}`pytask.Task`.
+- {pull}`235` refactors the utility functions for dealing with marks in
+  {mod}`_pytask.mark_utils`. (Closes {issue}`220`.)
 
 ## 0.1.9 - 2022-02-23
