+++ src/_pytask/persist.py
@@ -8,7 +8,7 @@ from _pytask.config import hookimpl
 from _pytask.dag import node_and_neighbors
 from _pytask.database import update_states_in_database
 from _pytask.exceptions import NodeNotFoundError
-from _pytask.mark_utils import get_specific_markers_from_task
+from _pytask.mark_utils import has_mark
 from _pytask.outcomes import Persisted
 from _pytask.outcomes import TaskOutcome
 
@@ -38,7 +38,7 @@ def pytask_execute_task_setup(session: Session, task: Task) -> None:
     The decorator needs to be set and all nodes need to exist.
 
     """
-    if get_specific_markers_from_task(task, "persist"):
+    if has_mark(task, "persist"):
         try:
             for name in node_and_neighbors(session.dag, task.name):
                 node = (
