+++ src/_pytask/resolve_dependencies.py
@@ -19,7 +19,8 @@ from _pytask.database import State
 from _pytask.exceptions import NodeNotFoundError
 from _pytask.exceptions import ResolvingDependenciesError
 from _pytask.mark import Mark
-from _pytask.mark_utils import get_specific_markers_from_task
+from _pytask.mark_utils import get_marks
+from _pytask.mark_utils import has_mark
 from _pytask.nodes import MetaNode
 from _pytask.nodes import Task
 from _pytask.path import find_common_ancestor_of_nodes
@@ -234,12 +235,12 @@ def _check_if_tasks_are_skipped(
 
 def _check_if_task_is_skipped(task_name: str, dag: nx.DiGraph) -> bool:
     task = dag.nodes[task_name]["task"]
-    is_skipped = get_specific_markers_from_task(task, "skip")
+    is_skipped = has_mark(task, "skip")
 
     if is_skipped:
         return True
 
-    skip_if_markers = get_specific_markers_from_task(task, "skipif")
+    skip_if_markers = get_marks(task, "skipif")
     is_any_true = any(
         _skipif(*marker.args, **marker.kwargs)[0] for marker in skip_if_markers
     )
