+++ src/pytask/__init__.py
@@ -12,10 +12,11 @@ from _pytask.mark import Mark
 from _pytask.mark import MARK_GEN as mark  # noqa: N811
 from _pytask.mark import MarkDecorator
 from _pytask.mark import MarkGenerator
-from _pytask.mark_utils import get_marks_from_obj
-from _pytask.mark_utils import get_specific_markers_from_task
-from _pytask.mark_utils import has_marker
-from _pytask.mark_utils import remove_markers_from_func
+from _pytask.mark_utils import get_all_marks
+from _pytask.mark_utils import get_marks
+from _pytask.mark_utils import has_mark
+from _pytask.mark_utils import remove_marks
+from _pytask.mark_utils import set_marks
 from _pytask.nodes import FilePathNode
 from _pytask.nodes import MetaNode
 from _pytask.nodes import Task
@@ -41,14 +42,15 @@ __all__ = [
     "cli",
     "console",
     "count_outcomes",
-    "get_marks_from_obj",
-    "get_specific_markers_from_task",
-    "has_marker",
+    "get_all_marks",
+    "get_marks",
+    "has_mark",
     "hookimpl",
     "import_optional_dependency",
     "main",
     "mark",
-    "remove_markers_from_func",
+    "remove_marks",
+    "set_marks",
     "CollectionOutcome",
     "Exit",
     "ExitCode",
