+++ tests/test_collect.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import itertools
 import sys
 import textwrap
 import warnings
@@ -7,6 +8,11 @@ from contextlib import ExitStack as does_not_raise  # noqa: N813
 from pathlib import Path
 
 import pytest
+from _pytask._collect import _check_that_names_are_not_used_multiple_times
+from _pytask._collect import _convert_to_dict
+from _pytask._collect import _merge_dictionaries
+from _pytask._collect import _Placeholder
+from _pytask._collect import convert_objects_to_node_dictionary
 from _pytask.collect import _find_shortest_uniquely_identifiable_name_for_tasks
 from _pytask.collect import pytask_collect_node
 from _pytask.exceptions import NodeNotCollectedError
@@ -233,4 +239,98 @@ def test_find_shortest_uniquely_identifiable_names_for_tasks(tmp_path):
         expected[task.name] = base_path.name + "/t.py::task_t"
 
     result = _find_shortest_uniquely_identifiable_name_for_tasks(tasks)
+    assert result == expected
+
+
+ERROR = "'@pytask.mark.depends_on' has nodes with the same name:"
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    ("x", "expectation"),
+    [
+        ([{0: "a"}, {0: "b"}], pytest.raises(ValueError, match=ERROR)),
+        ([{"a": 0}, {"a": 1}], pytest.raises(ValueError, match=ERROR)),
+        ([{"a": 0}, {"b": 0}, {"a": 1}], pytest.raises(ValueError, match=ERROR)),
+        ([{0: "a"}, {1: "a"}], does_not_raise()),
+        ([{"a": 0}, {0: "a"}], does_not_raise()),
+        ([{"a": 0}, {"b": 1}], does_not_raise()),
+    ],
+)
+def test_check_that_names_are_not_used_multiple_times(x, expectation):
+    with expectation:
+        _check_that_names_are_not_used_multiple_times(x, "depends_on")
+
+
+@pytest.mark.integration
+@pytest.mark.parametrize("when", ["depends_on", "produces"])
+@pytest.mark.parametrize(
+    "objects, expectation, expected",
+    [
+        ([0, 1], does_not_raise, {0: 0, 1: 1}),
+        ([{0: 0}, {1: 1}], does_not_raise, {0: 0, 1: 1}),
+        ([{0: 0}], does_not_raise, {0: 0}),
+        ([[0]], does_not_raise, {0: 0}),
+        (
+            [((0, 0),), ((0, 1),)],
+            does_not_raise,
+            {0: {0: 0, 1: 0}, 1: {0: 0, 1: 1}},
+        ),
+        ([{0: {0: {0: 0}}}, [2]], does_not_raise, {0: {0: {0: 0}}, 1: 2}),
+        ([{0: 0}, {0: 1}], ValueError, None),
+    ],
+)
+def testconvert_objects_to_node_dictionary(objects, when, expectation, expected):
+    expectation = (
+        pytest.raises(expectation, match=f"'@pytask.mark.{when}' has nodes")
+        if expectation == ValueError
+        else expectation()
+    )
+    with expectation:
+        nodes = convert_objects_to_node_dictionary(objects, when)
+        assert nodes == expected
+
+
+def _convert_placeholders_to_tuples(x):
+    counter = itertools.count()
+    return {
+        (next(counter), k.scalar)
+        if isinstance(k, _Placeholder)
+        else k: _convert_placeholders_to_tuples(v)
+        if isinstance(v, dict)
+        else v
+        for k, v in x.items()
+    }
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    "x, first_level, expected",
+    [
+        (1, True, {(0, True): 1}),
+        ({1: 0}, False, {1: 0}),
+        ({1: [2, 3]}, False, {1: {0: 2, 1: 3}}),
+        ([2, 3], True, {(0, False): 2, (1, False): 3}),
+        ([2, 3], False, {0: 2, 1: 3}),
+    ],
+)
+def test__convert_to_dict(x, first_level, expected):
+    """We convert placeholders to a tuple consisting of the key and the scalar bool."""
+    result = _convert_to_dict(x, first_level)
+    modified_result = _convert_placeholders_to_tuples(result)
+    assert modified_result == expected
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    "list_of_dicts, expected",
+    [
+        ([{1: 0}, {0: 1}], {1: 0, 0: 1}),
+        ([{_Placeholder(): 1}, {0: 0}], {1: 1, 0: 0}),
+        ([{_Placeholder(scalar=True): 1}], 1),
+        ([{_Placeholder(scalar=False): 1}], {0: 1}),
+    ],
+)
+def test_merge_dictionaries(list_of_dicts, expected):
+    result = _merge_dictionaries(list_of_dicts)
     assert result == expected
