+++ tests/test_nodes.py
@@ -1,19 +1,13 @@
 from __future__ import annotations
 
-import itertools
 from contextlib import ExitStack as does_not_raise  # noqa: N813
 from pathlib import Path
 
 import attr
 import pytask
 import pytest
-from _pytask.nodes import _check_that_names_are_not_used_multiple_times
-from _pytask.nodes import _convert_objects_to_node_dictionary
-from _pytask.nodes import _extract_nodes_from_function_markers
-from _pytask.nodes import _Placeholder
-from _pytask.nodes import convert_to_dict
+from _pytask.collect import _extract_nodes_from_function_markers
 from _pytask.nodes import depends_on
-from _pytask.nodes import merge_dictionaries
 from _pytask.nodes import produces
 from _pytask.shared import reduce_node_name
 from pytask import FilePathNode
@@ -88,26 +82,6 @@ def test_instantiation_of_metanode():
     assert isinstance(task, MetaNode)
 
 
-ERROR = "'@pytask.mark.depends_on' has nodes with the same name:"
-
-
-@pytest.mark.unit
-@pytest.mark.parametrize(
-    ("x", "expectation"),
-    [
-        ([{0: "a"}, {0: "b"}], pytest.raises(ValueError, match=ERROR)),
-        ([{"a": 0}, {"a": 1}], pytest.raises(ValueError, match=ERROR)),
-        ([{"a": 0}, {"b": 0}, {"a": 1}], pytest.raises(ValueError, match=ERROR)),
-        ([{0: "a"}, {1: "a"}], does_not_raise()),
-        ([{"a": 0}, {0: "a"}], does_not_raise()),
-        ([{"a": 0}, {"b": 1}], does_not_raise()),
-    ],
-)
-def test_check_that_names_are_not_used_multiple_times(x, expectation):
-    with expectation:
-        _check_that_names_are_not_used_multiple_times(x, "depends_on")
-
-
 @attr.s
 class FalseNode:
     path = attr.ib()
@@ -147,77 +121,3 @@ def test_reduce_node_name(node, paths, expectation, expected):
     with expectation:
         result = reduce_node_name(node, paths)
         assert result == expected
-
-
-@pytest.mark.integration
-@pytest.mark.parametrize("when", ["depends_on", "produces"])
-@pytest.mark.parametrize(
-    "objects, expectation, expected",
-    [
-        ([0, 1], does_not_raise, {0: 0, 1: 1}),
-        ([{0: 0}, {1: 1}], does_not_raise, {0: 0, 1: 1}),
-        ([{0: 0}], does_not_raise, {0: 0}),
-        ([[0]], does_not_raise, {0: 0}),
-        (
-            [((0, 0),), ((0, 1),)],
-            does_not_raise,
-            {0: {0: 0, 1: 0}, 1: {0: 0, 1: 1}},
-        ),
-        ([{0: {0: {0: 0}}}, [2]], does_not_raise, {0: {0: {0: 0}}, 1: 2}),
-        ([{0: 0}, {0: 1}], ValueError, None),
-    ],
-)
-def test_convert_objects_to_node_dictionary(objects, when, expectation, expected):
-    expectation = (
-        pytest.raises(expectation, match=f"'@pytask.mark.{when}' has nodes")
-        if expectation == ValueError
-        else expectation()
-    )
-    with expectation:
-        nodes = _convert_objects_to_node_dictionary(objects, when)
-        assert nodes == expected
-
-
-def _convert_placeholders_to_tuples(x):
-    counter = itertools.count()
-    return {
-        (next(counter), k.scalar)
-        if isinstance(k, _Placeholder)
-        else k: _convert_placeholders_to_tuples(v)
-        if isinstance(v, dict)
-        else v
-        for k, v in x.items()
-    }
-
-
-@pytest.mark.unit
-@pytest.mark.parametrize(
-    "x, first_level, expected",
-    [
-        (1, True, {(0, True): 1}),
-        ({1: 0}, False, {1: 0}),
-        ({1: [2, 3]}, False, {1: {0: 2, 1: 3}}),
-        ([2, 3], True, {(0, False): 2, (1, False): 3}),
-        ([2, 3], False, {0: 2, 1: 3}),
-    ],
-)
-def test_convert_to_dict(x, first_level, expected):
-    """We convert placeholders to a tuple consisting of the key and the scalar bool."""
-    result = convert_to_dict(x, first_level)
-    modified_result = _convert_placeholders_to_tuples(result)
-    assert modified_result == expected
-
-
-@pytest.mark.unit
-@pytest.mark.parametrize(
-    "list_of_dicts, expected",
-    [
-        ([{1: 0}, {0: 1}], {1: 0, 0: 1}),
-        ([{_Placeholder(): 1}, {0: 0}], {1: 1, 0: 0}),
-        ([{_Placeholder(scalar=True): 1}], 1),
-        ([{_Placeholder(scalar=False): 1}], {0: 1}),
-    ],
-)
-def test_merge_dictionaries(list_of_dicts, expected):
-    result = merge_dictionaries(list_of_dicts)
-    assert result == expected
